Hoist guts of sector read to ParseSectorInput
